Change type `blobs` in message Collation
From string to []byte

Additionally
- Isolate a test to `listeningshards_test.go`
- Refactor `Hash` to `keccak` and `GetCollationHash`

TODO
- Refactor the tests
- Debug issue in `AddPeer`
