+++ bootstrap.go
@@ -12,18 +12,6 @@ import (
 	ma "github.com/multiformats/go-multiaddr"
 )
 
-var (
-	IPFS_PEERS = convertPeers([]string{
-		"/ip4/127.0.0.1/tcp/4001/ipfs/QmXa1ncfGc9RQotUAyN3Gb4ar7WXZ4DEb2wPZsSybkK1sf",
-		"/ip4/127.0.0.1/tcp/4002/ipfs/QmbL37Cg7USSPPRu8uXtAL83dupVhyaHq4PrPhRAqxZTJ6", // our own node
-		"/ip4/104.131.131.82/tcp/4001/ipfs/QmaCpDMGvV2BGHeYERUEnRQAwe3N8SzbUtfsmvsqQLuvuJ",
-		"/ip4/104.236.179.241/tcp/4001/ipfs/QmSoLPppuBtQSGwKDZT2M73ULpjvfd3aZ6ha4oFGL1KrGM",
-		"/ip4/128.199.219.111/tcp/4001/ipfs/QmSoLSafTMBsPKadTEgaXctDQVcqN88CNLHXMkTNwMKPnu",
-		// "/ip4/104.236.76.40/tcp/4001/ipfs/QmSoLV4Bbm51jM9C4gDYZQ9Cy3U6aXMJDAbzgu2fzaDs64",
-		"/ip4/178.62.158.247/tcp/4001/ipfs/QmSoLer265NRgSp2LA3dPaeykiS1J6DifTC88f5uVQKNAd",
-	})
-)
-
 func convertPeers(peers []string) []pstore.PeerInfo {
 	pinfos := make([]pstore.PeerInfo, len(peers))
 	for i, peer := range peers {
