+++ pb/message/message.proto
@@ -29,7 +29,7 @@ message AddPeerResponse {
 message Collation {
     int64 shardID = 1;
     int64 period = 2;
-    string blobs = 3;
+    bytes blobs = 3;
 }
 
 message CollationRequest {
