+++ utils.go
@@ -5,13 +5,32 @@ import (
 	"fmt"
 	"log"
 
+	pbmsg "github.com/ethresearch/sharding-p2p-poc/pb/message"
 	inet "github.com/libp2p/go-libp2p-net"
 	peer "github.com/libp2p/go-libp2p-peer"
 	ma "github.com/multiformats/go-multiaddr"
 	protobufCodec "github.com/multiformats/go-multicodec/protobuf"
+	"golang.org/x/crypto/sha3"
 
 	"github.com/golang/protobuf/proto"
+	b58 "github.com/mr-tron/base58/base58"
 )
+
+func keccak(msg proto.Message) []byte {
+	dataInBytes, err := proto.Marshal(msg)
+	if err != nil {
+		log.Printf("Error occurs when hashing %v", msg)
+	}
+	h := sha3.NewLegacyKeccak256()
+	h.Write(dataInBytes)
+	return h.Sum(nil)
+}
+
+func GetCollationHash(msg *pbmsg.Collation) string {
+	hashBytes := keccak(msg)
+	// FIXME: for convenience
+	return b58.Encode(hashBytes)
+}
 
 // helper method - writes a protobuf go data object to a network stream
 // data: reference of protobuf go data object to send (not the object itself)
